requirements.txt
annotated-types==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
debugpy==1.6.7.post1 ; python_version >= "3.11" and python_version < "4.0"
django-debug-toolbar==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
django-polymorphic==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.5 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
drf-yasg==1.21.7 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
loguru==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.0.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2==2.9.7 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
